@@ -1,4 +1,4 @@
-import 'package:numbers_to_words/numbers_to_words.dart';
+import '../lib/numbers_to_words.dart';
 
 void main() {
   NumberToWords.convert(3, "da");
